Move the MetadataAsSourceFileService down to the features layer
None of this actually depended on the editor in any meaningful way.
