+++ src/EditorFeatures/CSharp/DecompiledSource/CSharpDecompiledSourceServiceFactory.cs
@@ -4,6 +4,7 @@
 
 using System;
 using System.Composition;
+using Microsoft.CodeAnalysis.DecompiledSource;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Host.Mef;
