+++ src/EditorFeatures/Core/Implementation/Peek/DefinitionPeekableItem.cs
@@ -6,6 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
+using Microsoft.CodeAnalysis.MetadataAsSource;
 using Microsoft.VisualStudio.Language.Intellisense;
 using Roslyn.Utilities;
