+++ src/Features/Core/Portable/DecompiledSource/IDecompiledSourceService.cs
@@ -6,7 +6,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Host;
 
-namespace Microsoft.CodeAnalysis.Editor
+namespace Microsoft.CodeAnalysis.DecompiledSource
 {
     internal interface IDecompiledSourceService : ILanguageService
     {
