+++ src/Features/Core/Portable/MetadataAsSource/MetadataAsSourceFile.cs
@@ -2,8 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-
-namespace Microsoft.CodeAnalysis.Editor
+namespace Microsoft.CodeAnalysis.MetadataAsSource
 {
     internal sealed class MetadataAsSourceFile
     {
