+++ src/Features/Core/Portable/MetadataAsSource/MetadataAsSourceGeneratedFileInfo.cs
@@ -11,7 +11,7 @@ using System.Text;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Text;
 
-namespace Microsoft.CodeAnalysis.Editor.Implementation.MetadataAsSource
+namespace Microsoft.CodeAnalysis.MetadataAsSource
 {
     internal sealed class MetadataAsSourceGeneratedFileInfo
     {
